refactor: upadate action names
